session15/requirements.txt
pillow
numpy
torch
torchvision
opencv-python
ipython
seaborn